+++ vsc-script/.vsc-script/rangeFinderTest.vsc-script.ts
@@ -1,19 +1,34 @@
-//import * as vsc from '../src/vsc-base-development/vsc-base'
-
 import * as ts from 'typescript'
 import * as vsc from 'vsc-base'
 import * as vscode from 'vscode'
+//import * as vsc from '../src/vsc-base-development/vsc-base'
 import { tsIsIdentifier } from '../src/vsc-base-development/vsc-base-typescript';
 
 let logger = '';
 
 
 export async function run(path: string) {
+	const fooo = () => { }
 	// let fileContent = await vsc.getFileContent(path);
 	// fileContent = fileContent.replace('let log =', 'let logger =')
 	// await vsc.saveFileContent(path, fileContent)
 	// await vsc.addFileContent(path, '//CONTENT IN THE END!!!')
 	// return;
+	const f = [
+		/**
+		 * asd
+		 */
+		//commetn11
+		//asdasd
+		fooo(),
+		/**
+		 * sdsss
+		 */
+		//commetn3245
+		//asd
+		fooo(),
+	]
+
 
 	vsc.showMessage("Start finding node... ")
 	const source = vsc.getDocumentContent()
@@ -21,7 +36,12 @@ export async function run(path: string) {
 		vsc.showMessage("No opnen document!")
 		return
 	}
+	const [node1, pos1] = vsc.tsFindNodePositionFromContent(source, node => ts.isCallExpression(node) && node.expression.getText() === 'fooo')
+	vsc.showMessage(pos1.content)
+	vsc.showMessage(pos1.fullContent)
+	vsc.setSelectionFromRange(pos1.range)
 
+	return
 
 	// let varFound = false
 	const [node, position] = vsc.tsFindNodePositionFromContent(source, node => {
@@ -103,4 +123,30 @@ export function method1(doit: boolean) {
 
 enum foo {
 	foo1 = 1 //<-- FIND this
+}
+
+{
+	"content": "foo33",
+		"start": 2521,
+			"end": 2526,
+				"startLineNumber": 86,
+					"endLineNumber": 86,
+						"startPosition": {
+		"line": 86,
+			"character": 10
+	},
+	"endPosition": {
+		"line": 86,
+			"character": 15
+	},
+	"range": [
+		{
+			"line": 86,
+			"character": 10
+		},
+		{
+			"line": 86,
+			"character": 15
+		}
+	]
 }